Separating Main from FlowControl
