+++ RSAkeyGen/src/com/borntocode/Main.java
@@ -1,0 +1,9 @@
+package com.borntocode;
+
+class Main {
+
+    public static void main(String[] args) {
+        FlowControl flowControl = new FlowControl();
+        flowControl.startMainLoop();
+    }
+}
